attached_assets/backend/requirements.txt

fastapi
uvicorn[standard]
requests
PyJWT>=2.8.0
cryptography
pydantic
